refactor(deploy): give the certification protocol its own module, and release database handles
Two prerequisites for moving certification into a helper process, both worth
having on their own.

The verdict constants, progress slots and `describeProgress` move to
`components/certificationProtocol.ts`, which imports nothing from Harper. Both
ends of the protocol need them, and one end is about to become a process whose
whole job is to report a verdict even when the module graph it is testing does not
load — so it must not reach `certifyCandidate` (and through it `manageThreads`) to
find out what a verdict looks like. `certifyCandidate` re-exports them so existing
importers are unaffected.

The module also defines the helper's exit codes now, because a SharedArrayBuffer
cannot cross a process boundary: the flag the candidate's thread writes stays
readable only inside the helper, which encodes the answer as its exit status. That
survives a lost IPC message the way the flag survives a lost worker message, and a
candidate cannot set it — worker threads have no `process.send`, and nothing in the
candidate's thread chooses the helper's exit. Any other code is a rejection, so
silence still cannot become a pass.

Second, the validator now calls `closeLoadedDatabases()` in its teardown, on pass
and on rejection alike. It reaches `getTables()` through `loadRootPlugins`, which
opens the whole database graph, and `resources/databases.ts` documents that a
thread exiting without closing leaks process-global RocksDB handles and blocks an
online `restore_backup` from confirming a database is closed. `jobProcess.ts`
already does this; the validator did not, so every certification leaked. Found by
the planning review. It is attempted independently of the scope closes so one
failure cannot skip the other, and its own failure is logged rather than allowed to
mask the load result.

Verified: 10 certification unit tests pass. The leak itself is asserted by reading
— registry refcount coverage is listed in the design note's verification route and
belongs with the helper-process change, where forced death is the backstop.
